+++ src/seppl/__init__.py
@@ -8,5 +8,5 @@ from ._class_registry import ClassListerRegistry, get_class_lister, DEFAULT
 from ._entry_points import generate_entry_points
 from ._help import generate_plugin_usage, generate_help, HELP_FORMATS, HELP_FORMAT_TEXT, HELP_FORMAT_MARKDOWN
 from ._types import get_class, get_class_name, classes_to_str, AnyData
-from ._metadata import MetaDataHandler, add_metadata, get_metadata
+from ._metadata import MetaDataHandler, add_metadata, get_metadata, METADATA_TYPES, METADATA_TYPE_STRING, METADATA_TYPE_NUMERIC, METADATA_TYPE_BOOL
 from ._session import Session, SessionHandler