+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - added `stopped` flag to `Session` to indicate that the execution is over,
   which is monitored by the `seppl.io.execute(...)` function
 - added `wai_logging>=0.0.5` as dependency
+- added constants for meta-data types: METADATA_TYPE_STRING, METADATA_TYPE_BOOL, METADATA_TYPE_NUMERIC
 
 
 0.2.21 (2025-08-19)